refactor(sql): share lexical scanning helpers
